+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 ## [Unreleased]
+- [CM-977] Fixed Typing Indicator being shown when user opens the older conversation
 - [CM-979] Fixed Last button of Curved/Quick Reply Button not supporting multiple lines 
 ## [0.2.3] - 2022-06-23
 - Upgraded KM Core SDK to 1.0.4